chore(txn-table): better name for the variables
