+++ apps/storybook/src/stories/transactions/TransactionTable.stories.ts
@@ -6,51 +6,51 @@ import { withWagmiProvider } from "../decorators/wagmi";
 import { TransactionTable } from "./TransactionTable";
 
 const meta = {
-	title: "Transactions/TransactionTable",
-	component: TransactionTable,
-	args: {},
-	parameters: {
-		layout: "centered",
-	},
-	decorators: [withWagmiProvider()],
-	argTypes: {
-		explorer: { options: Object.values(Explorer) },
-	},
+  title: "Transactions/TransactionTable",
+  component: TransactionTable,
+  args: {},
+  parameters: {
+    layout: "centered",
+  },
+  decorators: [withWagmiProvider()],
+  argTypes: {
+    explorer: { options: Object.values(Explorer) },
+  },
 } satisfies Meta<typeof TransactionTable>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
-export const TransactionTableWithoutFetching: Story = {
-	args: {
-		transactions: TXN_LIST.map((txn) => asTransactionMeta(txn)),
-	},
+export const TransactionTableSome: Story = {
+  args: {
+    transactions: TXN_LIST.map((txn) => asTransactionMeta(txn)),
+  },
 };
 
-export const TransactionTableWithoutFetching2: Story = {
-	args: {
-		transactions: generateTxnFixturesByCount(50).map((txn) =>
-			asTransactionMeta(txn),
-		),
-	},
+export const TransactionTableMany: Story = {
+  args: {
+    transactions: generateTxnFixturesByCount(50).map((txn) =>
+      asTransactionMeta(txn),
+    ),
+  },
 };
 
 export const TransactionTableWithBlockscoutExplorerLink: Story = {
-	args: {
-		transactions: generateTxnFixturesByCount(50).map((txn) =>
-			asTransactionMeta(txn),
-		),
-		explorer: Explorer.Blockscout,
-		chainId: 1,
-	},
+  args: {
+    transactions: generateTxnFixturesByCount(50).map((txn) =>
+      asTransactionMeta(txn),
+    ),
+    explorer: Explorer.Blockscout,
+    chainId: 1,
+  },
 };
 
 export const TransactionTableWithBlockscoutExplorerLinkInOptimism: Story = {
-	args: {
-		transactions: generateTxnFixturesByCount(50).map((txn) =>
-			asTransactionMeta(txn),
-		),
-		explorer: Explorer.Blockscout,
-		chainId: 10,
-	},
+  args: {
+    transactions: generateTxnFixturesByCount(50).map((txn) =>
+      asTransactionMeta(txn),
+    ),
+    explorer: Explorer.Blockscout,
+    chainId: 10,
+  },
 };
